+++ services/backend/src/main/java/com/soprasteria/g4it/backend/apievaluating/business/asyncevaluatingservice/EvaluateAiService.java
@@ -396,7 +396,7 @@ public class EvaluateAiService {
                             .add(values);
 
                     if (evaluateReportBO.isExport()) {
-                        csvOutVirtualEquipment.printRecord(impactToCsvRecord.toCsv(
+                        csvOutVirtualEquipment.printRecord(impactToCsvRecord.toCsvAi(
                                 context, evaluateReportBO, virtualEquipment, impact, refSip.get(impact.getCritere()))
                         );
                     }