+++ services/frontend/src/app/layout/digital-services-footprint/digital-services-footprint-dashboard/digital-services-footprint-dashboard.component.ts
@@ -94,10 +94,9 @@ export class DigitalServicesFootprintDashboardComponent
     digitalService: DigitalService = {} as DigitalService;
     aiRecommendation: AiRecommendation = {} as AiRecommendation;
     showDataButton = false;
-    displaySetViewPopup = false;
     title = "";
     content = "";
-
+    isCollapsed = false;
     impacts: any[] = [];
     topThreeImpacts: any[] = [];
     topPieThreeImpacts: any[] = [];
@@ -112,6 +111,7 @@ export class DigitalServicesFootprintDashboardComponent
     displayCriteriaPopup = false;
     workspace: WorkspaceWithOrganization = {} as WorkspaceWithOrganization;
     organization!: Organization;
+    filterSidebarVisible = false;
     textDescriptionImpacts: {
         text: string;
         impactName: string;
@@ -294,7 +294,7 @@ export class DigitalServicesFootprintDashboardComponent
         }
     }
 
-    handleChartChange(criteria: string) {
+    handleChartChange(criteria: any) {
         if (this.selectedCriteria === "Global Vision") {
             this.selectedCriteria = criteria;
             this.chartType.set("pie");
@@ -682,6 +682,11 @@ export class DigitalServicesFootprintDashboardComponent
                 this.barChartChild = true;
             }
         }
+    }
+
+    handleFilters(event: { enableConsistency: boolean; unitType: string }) {
+        this.selectedUnit = event.unitType;
+        this.updateDataConsistencyInDS(event.enableConsistency);
     }
 
     ngOnDestroy() {